Variable screen size
This fucking- this little bit of code literally just,
it keeps the surface area of the screen (in in-game pixels) the same across multiple screen sizes and aspect ratios.
It took bloody six months. This was one of the first things I tried doing when I started on the engine before it was even called Glass. Before it even drew to the canvas. I just wanted consistent sizing. Now I have it. I'm so happy.